requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Azure SDKs
azure-ai-formrecognizer==3.3.0
azure-ai-textanalytics==5.3.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# MySQL Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Data processing
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
PyPDF2==3.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1